Including `sdkconfig` - *just in case!*
...Hope I didn't infringe any copyrights with ESP-IDF components. It's all FLOSS, usually.

(And that `managed_components` directory was HUGE. 100 MiB+ in size. How did it get through so quickly?)
PS The reason why I committed all of this stuff was that I had to modify a few files to actually get this project to build.
These ones. From Espressif's Arduino implementation, I think:
- `AP.cpp`,
- `STA.cpp`,
- `WiFiGeneric.cpp`,
- `chip-debug-report.cpp`.